Determines whether or not to create a new thumb
Instead of relying on the existence of the cached thumbnail to be the only indication of whether or not a thumbnail should be created, check if the thumbnail exists and if it does check the modified time of the source file and compare it to the modified time of the cached thumbnail. If the source file is newer than the thumbnail, or there is no pre-existing thumbnail, generate the thumbnail for the source.